A better boot script:  - Don't download vmlinuz/initrd unless needed  - Don't recreate disk files unless needed  - Console instead of graphic output

+++ boot
@@ -4,19 +4,29 @@ set -e
 
 source ./config
 
-#wget $VMLINUZ_URL
-#wget $INITRD_URL
-#echo $KS_FILE | cpio -c -o >> initrd.img
+wget -N $VMLINUZ_URL
+wget -N $INITRD_URL
 
-dd if=/dev/zero of=system.raw bs=1024k count=$SYSTEM_DISK_SIZE
-dd if=/dev/zero of=data.raw bs=1024k count=$DATA_DISK_SIZE
+cp initrd.img initrd_ks.img
+echo $KS_FILE | cpio -c -o >> initrd_ks.img
 
+if [ ! -f system.raw ]; then
+    dd if=/dev/zero of=system.raw bs=1024k count=$SYSTEM_DISK_SIZE
+fi
 
-sudo qemu-system-x86_64 -hda system.raw -hdb data.raw -net nic -net user -m $MEMORY_SIZE -localtime -enable-kvm -kernel vmlinuz -initrd initrd.img -append "ks=file:///$KS_FILE" #-nographic
+if [ ! -f data.raw ]; then
+    dd if=/dev/zero of=data.raw bs=1024k count=$DATA_DISK_SIZE
+fi
 
-exit
-qemu-system-x86_64
-    -cdrom Fedora-Live-Desktop-x86_64-20-1.iso  \
-    -hda system.raw                             \
-    -hdb data.raw                               \
-    -nographic
+sudo qemu-system-x86_64     \
+        -hda system.raw     \
+        -hdb data.raw       \
+        -net nic -net user  \
+        -m $MEMORY_SIZE     \
+        -localtime          \
+        -enable-kvm         \
+        -nographic          \
+        -no-reboot          \
+        -kernel vmlinuz     \
+        -initrd initrd_ks.img   \
+        -append "ks=file:///$KS_FILE console=ttyS0 panic=1"
